autoupdate: rewrite command
It was originally written as a Homebrew internal command, but we decided not
to ship it at that point in time so I moved it over to supporting it myself
for now, as there was reasonable demand for it and I use it personally.

Now that it seems to be gaining usage it seemed a good idea to invest the time
to rewrite the command properly, including stamping out some implementation hacks
and making some style tweaks to keep Rubocop happy.

Has been pretty thoroughly tested locally but please do shout if you run into
any issues locally.