autoupdate: stop abusing plists
Whilst the previous version worked it also was kind of a hack and made
plutil deeply unhappy, which some third-party tools use to check for
validity in terms of loading or not loading the plist at launch as
intended.

There's a very reasonable argument that you shouldn't use plists in the
way I was doing so, because they weren't really designed for that.
Consequently, I've gone with the approach a decent amount of tools do
and placed a temporary script inside the user cache directory which
handles the actual command execution. The plist now only exists
to trigger that executable.

Closes https://github.com/DomT4/homebrew-autoupdate/issues/6.
Closes https://github.com/DomT4/homebrew-autoupdate/pull/8.