autoupdate: add notification support
I decided to use terminal-notifier instead of osascript because
osascript has an enormously annoying habit of opening the script
editor app if you click the notification, which people understandably
are quite prone to doing. If anyone knows a good way around that please
let me know.

Since I'm not vendoring terminal-notifier into this project to keep it
slim, this does require you to have installed it either through brew
or to have installed the `.app` bundle. Either will work. It also
requires you to be running no earlier than OS X/macOS 10.8, which these
days feels like a reasonable request.

I might go in a different direction with this in future, but I think
this is the best way to get the ball rolling on a simple nudge that
autoupdate has successfully completed. There isn't yet a start
notification because I'm unsure on the merits if there's an end
notification. But I'll think about it some more, depending on feedback.

The notification also doesn't do anything at this point in terms of
opening logs when clicked, etc. That may also change, depending on
feedback. I was considering adding a `--show-logs` command which would
essentially cat the two logs & provide a terminal output for people to
glance over, but I'm wary of adding more moving parts to this prior
to there being enough demand to merit the potential for breakage.

Hopefully this is a decent enough start for people to feel happy though.
I think it's a minimal enough place to begin to not be too disruptive.

Closes https://github.com/DomT4/homebrew-autoupdate/issues/5.