start: drop notification support for < Yosemite
terminal-notifier upstream has deprecated 10.8 & 10.9 in the
latest release & I don't have a wide enough testing system to
certify I'm not making breaking changes over time on the
notification side of things, so for now I'm implementing a block
here.

I'll reconsider options if there's enough of an outcry, but
I know from my time as a Homebrew maintainer usage of Mavericks
and below is a pretty small fraction of users.