start: warn of terminal-notifier support deprecation
From October 2022 (ie the next version of macOS) support for
terminal-notifier will be removed. This effectively will remove
notificiation support from macOS versions prior to Catalina/10.15.

This is due to `terminal-notifier` essentially being abandonware
upstream and being increasingly difficult to test on newer versions
of macOS to ensure it still actually functions as intended here.

The expected impact is minimal. The rest of this tap, minus notification support,
will continue functioning as normal for older versions of macOS.